part5/requirements.txt
cryptography~=42.0.5
apache-airflow~=2.8.1
apache-airflow-providers-postgres~=5.10.0
numpy~=1.24.4
pandas~=2.0.3
psycopg2-binary~=2.9.7
requests~=2.31.0
tenacity~=8.2.3